requirements.txt
atlassian-python-api
python-dotenv
mcp
